update menu button styling
