+++ packages/nodes/src/views/document-view/components/context-menu-container/common/MenuButton.tsx
@@ -9,14 +9,14 @@ type MenuButtonProps = {
 export const MenuButton = ({ icon, label, action }: MenuButtonProps) => {
     return (
         <button
-            className='np-w-full np-p-2 np-mt-1 np-mb-1 first:np-mt-0 last:np-mb-0 np-flex np-items-center np-rounded-md np-bg-light hover:np-bg-grey'
+            className='np-w-full np-h-8 np-pl-[6px] np-mb-1 last:np-mb-0 np-flex np-items-center np-rounded-sm np-bg-light hover:np-bg-grey'
             onClick={action}
         >
-            <div className='np-w-6 np-h-6 np-flex np-items-center np-justify-center'>
+            <div className='np-h-full np-w-5 np-flex np-flex-col np-items-center np-justify-center'>
                 {icon}
             </div>
-            <div className='np-flex np-flex-grow np-items-center np-justify-start np-pl-4'>
-                <p className='np-font-medium np-font-sans np-text-dark'>{label}</p>
+            <div className='np-pl-2 np-flex np-flex-grow np-items-center np-justify-start'>
+                <p className='np-font-sans np-font-medium np-text-dark np-text-sm -np-translate-y-px'>{label}</p>
             </div>
         </button>
     )
