+++ packages/nodes/src/views/document-view/components/context-menu-container/menus/index.ts
@@ -1,1 +1,1 @@
-export { PortContextMenu } from './PortContextMenu'
+export { PortContextMenu } from './port-context-menu'
