+++ packages/nodes/src/views/document-view/components/context-menu-container/menus/port-context-menu/buttons/index.ts
@@ -1,0 +1,1 @@
+export { PinButton } from './PinButton'
