+++ packages/nodes/src/views/document-view/components/context-menu-container/menus/port-context-menu/index.ts
@@ -1,0 +1,3 @@
+import PortContextMenu from './PortContextMenu'
+
+export { PortContextMenu }
