+++ packages/nodes/src/views/document-view/components/context-menu-container/utils/getMenuHeight.ts
@@ -4,7 +4,7 @@ export const getMenuHeight = (context: ContextMenuContext): number => {
     switch (context.type) {
         case 'port': {
             const { direction } = context
-            return direction === 'input' ? 100 : 40
+            return direction === 'input' ? 220 : 40
         }
         default: {
             return 5
